backend/pyproject.toml
[tool.poetry]
name = "dom-enterprise-gateway-backend"
version = "0.1.0"
description = "Backend for DOM Enterprise Gateway"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
langchain = ">=1.1.0,<2.0.0"
langchain-postgres = ">=0.0.16,<0.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
alembic = "^1.13.1"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
redis = "^5.0.5"
authlib = "^1.3.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
asyncpg = ">=0.30.0,<1.0.0"
langchain-google-genai = ">=3.2.0,<4.0.0"
email-validator = "^2.2.0" 
aiosqlite = "^0.20.0"
pydantic-settings = "^2.6.0"   # ★ この行を追加

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
